moved draw logic for noise halo to other file
