+++ shaders/utils.py
@@ -95,38 +95,3 @@ def light1(intensity, attenuation, dist):
 @ti.func
 def light2(intensity, attenuation, dist):
     return intensity / (1.0 + dist * dist * attenuation)
-
-
-@ti.func
-def draw(vuv, time):
-    uv = vuv
-    ang = ts.atan(uv.y, uv.x)
-    uv_len = length(uv)
-    color1 = ts.vec3(0.611765, 0.262745, 0.996078)
-    color2 = ts.vec3(0.298039, 0.760784, 0.913725)
-    color3 = ts.vec3(0.062745, 0.078431, 0.600000)
-    inner_radius = 0.6
-    noise_scale = 0.65
-    
-    n0 = s_noise_3(ts.vec3(uv.x * noise_scale, uv.y * noise_scale, time * 0.5)) * 0.5 + 0.5
-    r0 = ts.mix(ts.mix(inner_radius, 1.0, 0.4), ts.mix(inner_radius, 1.0, 0.6), n0)
-    d0 = ts.distance(uv, r0 / uv_len * uv)
-    v0 = light1(1.0, 10.0, d0)
-    v0 *= smoothstep(r0 * 1.05, r0, uv_len)
-    cl = ti.cos(ang + time * 2.0) * 0.5 + 0.5
-    
-    a = time * (-1.0)
-    pos = ts.vec2(ti.cos(a), ti.sin(a)) * r0
-    d = ts.distance(uv, pos)
-    v1 = light2(1.5, 5.0, d)
-    v1 *= light1(1.0, 50.0, d0)
-    
-    v2 = smoothstep(1.0, ts.mix(inner_radius, 1.0, n0 * 0.5), uv_len)
-    v3 = smoothstep(inner_radius, ts.mix(inner_radius, 1.0, 0.5), uv_len)
-    
-    col = ts.mix(color1, color2, cl)
-    col = ts.mix(color3, col, v0)
-    col = (col + v1) * v2 * v3
-    
-    col.rgb = clamp(col.rgb, 0.0, 1.)
-    return extract_alpha(col)
